+++ mobilecoind-dev-faucet/Cargo.toml
@@ -17,14 +17,13 @@ mc-mobilecoind-api = { path = "../mobilecoind/api" }
 mc-transaction-core = { path = "../transaction/core" }
 mc-util-grpc = { path = "../util/grpc" }
 mc-util-keyfile = { path = "../util/keyfile" }
+mc-util-serial = { path = "../util/serial" }
 
 clap = { version = "3.1", features = ["derive", "env"] }
 grpcio = "0.10.2"
 hex = "0.4"
-protobuf = "2.27.1"
 rocket = { version = "0.5.0-rc.2", features = ["json"] }
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
-serde_with = "1.13"
 tokio = "1"
