+++ mobilecoind-json/Cargo.toml
@@ -21,7 +21,6 @@ protobuf = "2.27.1"
 rocket = { version = "0.5.0-rc.2", features = ["json"] }
 serde = "1.0"
 serde_derive = "1.0"
-serde_with = "1.13"
 
 [dev-dependencies]
 mc-crypto-keys = { path = "../crypto/keys" }
@@ -29,5 +28,6 @@ mc-ledger-db = { path = "../ledger/db" }
 mc-transaction-core = { path = "../transaction/core" }
 mc-transaction-core-test-utils = { path = "../transaction/core/test-utils" }
 mc-util-from-random = { path = "../util/from-random" }
+mc-util-serial = { path = "../util/serial" }
 
 rand = "0.8"
